test: share stack frame util
